Environment_Monitor/Device/requirements.txt
azure-iot-device
Adafruit_DHT